+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.0.3 » Future release
+#### New Features
+* Added parameter `guild` for methods `DiscordLevelingSystem.is_in_database()` and `DiscordLevelingSystem.remove_from_database()` This allows a more targeted check or removal for the specified member 
+  * `DiscordLevelingSystem.is_in_database(member: Union[Member, int], guild: Guild=None)`
+  * `DiscordLevelingSystem.remove_from_database(member: Union[Member, int], guild: Guild=None)`
 ## v0.0.2 » Jun. 21, 2021
 #### New Features
 * Added the ability for `LevelUpAnnouncement` messages to be embeds ([docs](https://github.com/Defxult/discordLevelingSystem#levelupannouncement))